+++ MonsterTerra(Capstone)/Assets/Scenes/Main.unity
@@ -10001,7 +10001,7 @@ GameObject:
   m_Component:
   - component: {fileID: 237992291}
   - component: {fileID: 237992293}
-  - component: {fileID: 237992292}
+  - component: {fileID: 237992294}
   m_Layer: 5
   m_Name: HP
   m_TagString: Untagged
@@ -10030,19 +10030,6 @@ RectTransform:
   m_AnchoredPosition: {x: -99.5, y: -5}
   m_SizeDelta: {x: -201, y: 7}
   m_Pivot: {x: 0.5, y: 1}
---- !u!114 &237992292
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 237992290}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b5b9d523788e375459995041e7e321bd, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  type: 4
 --- !u!114 &237992293
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -10094,6 +10081,19 @@ MonoBehaviour:
   m_OnValueChanged:
     m_PersistentCalls:
       m_Calls: []
+--- !u!114 &237992294
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 237992290}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: db0c5d8f92bcb42488aef4a7bca9938c, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  type: 4
 --- !u!1 &238604516
 GameObject:
   m_ObjectHideFlags: 0
@@ -10447,7 +10447,7 @@ GameObject:
   m_Component:
   - component: {fileID: 407549671}
   - component: {fileID: 407549673}
-  - component: {fileID: 407549672}
+  - component: {fileID: 407549674}
   m_Layer: 5
   m_Name: MP
   m_TagString: Untagged
@@ -10476,19 +10476,6 @@ RectTransform:
   m_AnchoredPosition: {x: -99.5, y: -12}
   m_SizeDelta: {x: -201, y: 7}
   m_Pivot: {x: 0.5, y: 1}
---- !u!114 &407549672
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 407549670}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b5b9d523788e375459995041e7e321bd, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  type: 5
 --- !u!114 &407549673
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -10540,6 +10527,19 @@ MonoBehaviour:
   m_OnValueChanged:
     m_PersistentCalls:
       m_Calls: []
+--- !u!114 &407549674
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 407549670}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: db0c5d8f92bcb42488aef4a7bca9938c, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  type: 5
 --- !u!1 &410525174
 GameObject:
   m_ObjectHideFlags: 0
@@ -14518,6 +14518,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: db0c5d8f92bcb42488aef4a7bca9938c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  type: 1
 --- !u!4 &1140202848
 Transform:
   m_ObjectHideFlags: 0
@@ -20415,7 +20416,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1683565384}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b5b9d523788e375459995041e7e321bd, type: 3}
+  m_Script: {fileID: 11500000, guid: db0c5d8f92bcb42488aef4a7bca9938c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   type: 6
@@ -21459,7 +21460,7 @@ GameObject:
   m_Component:
   - component: {fileID: 1894611448}
   - component: {fileID: 1894611450}
-  - component: {fileID: 1894611449}
+  - component: {fileID: 1894611451}
   m_Layer: 5
   m_Name: EXP
   m_TagString: Untagged
@@ -21491,19 +21492,6 @@ RectTransform:
   m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 0, y: 5}
   m_Pivot: {x: 0.5, y: 1}
---- !u!114 &1894611449
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1894611447}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b5b9d523788e375459995041e7e321bd, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  type: 0
 --- !u!114 &1894611450
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -21555,6 +21543,19 @@ MonoBehaviour:
   m_OnValueChanged:
     m_PersistentCalls:
       m_Calls: []
+--- !u!114 &1894611451
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1894611447}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: db0c5d8f92bcb42488aef4a7bca9938c, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  type: 0
 --- !u!1 &1905244759
 GameObject:
   m_ObjectHideFlags: 0
