--- MonsterTerra(Capstone)/Assets/Script/Player/HUD.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b5b9d523788e375459995041e7e321bd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
